311/requirements.txt
redis==5.0.3
kafka-python==2.0.2